rename cirucit/primitives.rs to circuit/initial_values.rs
