+++ crates/light-client-prover/src/circuit/initial_values.rs
